chore: Remove process function refactor
