+++ go/packages/dev-mode/lib/logger.go
@@ -81,7 +81,7 @@ func ReportLog(logPayload string) {
 	_, ok := os.LookupEnv("SLS_DEBUG_EXTENSION")
 	_, toLogs := os.LookupEnv("SLS_TEST_EXTENSION_LOG")
 	if ok || toLogs {
-		BaseLogger.Printf("⚡ DEV-MODE: Log###%s", base64.StdEncoding.EncodeToString([]byte(logPayload)))
+		BaseLogger.Printf("SERVERLESS_TELEMETRY.DL.%s", base64.StdEncoding.EncodeToString([]byte(logPayload)))
 	}
 }
 
@@ -89,7 +89,7 @@ func ReportReqRes(logPayload string) {
 	_, ok := os.LookupEnv("SLS_DEBUG_EXTENSION")
 	_, toLogs := os.LookupEnv("SLS_TEST_EXTENSION_LOG")
 	if ok || toLogs {
-		BaseLogger.Printf("⚡ DEV-MODE: ReqRes###%s", base64.StdEncoding.EncodeToString([]byte(logPayload)))
+		BaseLogger.Printf("SERVERLESS_TELEMETRY.DR.%s", base64.StdEncoding.EncodeToString([]byte(logPayload)))
 	}
 }
 
@@ -97,7 +97,7 @@ func ReportSpans(logPayload string) {
 	_, ok := os.LookupEnv("SLS_DEBUG_EXTENSION")
 	_, toLogs := os.LookupEnv("SLS_TEST_EXTENSION_LOG")
 	if ok || toLogs {
-		BaseLogger.Printf("⚡ DEV-MODE: Traces###%s", base64.StdEncoding.EncodeToString([]byte(logPayload)))
+		BaseLogger.Printf("SERVERLESS_TELEMETRY.DM.%s", base64.StdEncoding.EncodeToString([]byte(logPayload)))
 	}
 }
